requirement.txt
streamlit
requests
matplotlib
python-dotenv
newsapi-python
eventregistry